use BaseConfigDatabaseTest
