+++ airbyte-config/config-persistence/src/test/java/io/airbyte/config/persistence/DatabaseConfigPersistenceTest.java
@@ -8,6 +8,8 @@ import static io.airbyte.config.ConfigSchema.STANDARD_DESTINATION_DEFINITION;
 import static io.airbyte.config.ConfigSchema.STANDARD_SOURCE_DEFINITION;
 import static io.airbyte.db.instance.configs.jooq.generated.Tables.ACTOR_DEFINITION;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.jooq.impl.DSL.asterisk;
+import static org.jooq.impl.DSL.count;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -18,19 +20,17 @@ import static org.mockito.Mockito.spy;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.Lists;
 import io.airbyte.commons.json.Jsons;
+import io.airbyte.config.ConfigSchema;
 import io.airbyte.config.ConfigWithMetadata;
+import io.airbyte.config.DestinationConnection;
+import io.airbyte.config.Geography;
+import io.airbyte.config.SourceConnection;
 import io.airbyte.config.StandardDestinationDefinition;
 import io.airbyte.config.StandardSourceDefinition;
 import io.airbyte.config.StandardSourceDefinition.ReleaseStage;
+import io.airbyte.config.StandardSourceDefinition.SourceType;
+import io.airbyte.config.StandardWorkspace;
 import io.airbyte.config.persistence.DatabaseConfigPersistence.ConnectorInfo;
-import io.airbyte.db.factory.DSLContextFactory;
-import io.airbyte.db.factory.DataSourceFactory;
-import io.airbyte.db.factory.FlywayFactory;
-import io.airbyte.db.instance.configs.ConfigsDatabaseMigrator;
-import io.airbyte.db.instance.configs.ConfigsDatabaseTestProvider;
-import io.airbyte.db.instance.development.DevDatabaseMigrator;
-import io.airbyte.db.instance.development.MigrationDevHelper;
-import io.airbyte.test.utils.DatabaseConnectionHelper;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDate;
@@ -38,10 +38,15 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Set;
 import java.util.UUID;
-import org.jooq.SQLDialect;
-import org.junit.jupiter.api.AfterEach;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import org.jooq.Record1;
+import org.jooq.Result;
+import org.jooq.Table;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -50,27 +55,73 @@ import org.junit.jupiter.api.Test;
  * methods.
  */
 @SuppressWarnings({"PMD.SignatureDeclareThrowsException", "PMD.ShortVariable", "PMD.JUnitTestsShouldIncludeAssert"})
-class DatabaseConfigPersistenceTest extends BaseDatabaseConfigPersistenceTest {
+class DatabaseConfigPersistenceTest extends BaseConfigDatabaseTest {
+
+  public static final String DEFAULT_PROTOCOL_VERSION = "0.2.0";
+  protected static final StandardSourceDefinition SOURCE_GITHUB = new StandardSourceDefinition()
+      .withName("GitHub")
+      .withSourceDefinitionId(UUID.fromString("ef69ef6e-aa7f-4af1-a01d-ef775033524e"))
+      .withDockerRepository("airbyte/source-github")
+      .withDockerImageTag("0.2.3")
+      .withDocumentationUrl("https://docs.airbyte.io/integrations/sources/github")
+      .withIcon("github.svg")
+      .withSourceType(SourceType.API)
+      .withProtocolVersion(DEFAULT_PROTOCOL_VERSION)
+      .withTombstone(false);
+  protected static final StandardSourceDefinition SOURCE_POSTGRES = new StandardSourceDefinition()
+      .withName("Postgres")
+      .withSourceDefinitionId(UUID.fromString("decd338e-5647-4c0b-adf4-da0e75f5a750"))
+      .withDockerRepository("airbyte/source-postgres")
+      .withDockerImageTag("0.3.11")
+      .withDocumentationUrl("https://docs.airbyte.io/integrations/sources/postgres")
+      .withIcon("postgresql.svg")
+      .withSourceType(SourceType.DATABASE)
+      .withTombstone(false);
+  protected static final StandardSourceDefinition SOURCE_CUSTOM = new StandardSourceDefinition()
+      .withName("Custom")
+      .withSourceDefinitionId(UUID.fromString("baba338e-5647-4c0b-adf4-da0e75f5a750"))
+      .withDockerRepository("airbyte/cusom")
+      .withDockerImageTag("0.3.11")
+      .withDocumentationUrl("https://docs.airbyte.io/integrations/sources/postgres")
+      .withIcon("postgresql.svg")
+      .withSourceType(SourceType.DATABASE)
+      .withCustom(true)
+      .withReleaseStage(ReleaseStage.CUSTOM)
+      .withTombstone(false);
+  protected static final StandardDestinationDefinition DESTINATION_SNOWFLAKE = new StandardDestinationDefinition()
+      .withName("Snowflake")
+      .withDestinationDefinitionId(UUID.fromString("424892c4-daac-4491-b35d-c6688ba547ba"))
+      .withDockerRepository("airbyte/destination-snowflake")
+      .withDockerImageTag("0.3.16")
+      .withDocumentationUrl("https://docs.airbyte.io/integrations/destinations/snowflake")
+      .withProtocolVersion(DEFAULT_PROTOCOL_VERSION)
+      .withTombstone(false);
+  protected static final StandardDestinationDefinition DESTINATION_S3 = new StandardDestinationDefinition()
+      .withName("S3")
+      .withDestinationDefinitionId(UUID.fromString("4816b78f-1489-44c1-9060-4b19d5fa9362"))
+      .withDockerRepository("airbyte/destination-s3")
+      .withDockerImageTag("0.1.12")
+      .withDocumentationUrl("https://docs.airbyte.io/integrations/destinations/s3")
+      .withProtocolVersion(DEFAULT_PROTOCOL_VERSION)
+      .withTombstone(false);
+  protected static final StandardDestinationDefinition DESTINATION_CUSTOM = new StandardDestinationDefinition()
+      .withName("Custom")
+      .withDestinationDefinitionId(UUID.fromString("baba338e-5647-4c0b-adf4-da0e75f5a750"))
+      .withDockerRepository("airbyte/cusom")
+      .withDockerImageTag("0.3.11")
+      .withDocumentationUrl("https://docs.airbyte.io/integrations/sources/postgres")
+      .withIcon("postgresql.svg")
+      .withCustom(true)
+      .withReleaseStage(StandardDestinationDefinition.ReleaseStage.CUSTOM)
+      .withTombstone(false);
+  private static final String CANNOT_BE_NULL = "can not be null";
+
+  private DatabaseConfigPersistence configPersistence;
 
   @BeforeEach
   public void setup() throws Exception {
-    dataSource = DatabaseConnectionHelper.createDataSource(container);
-    dslContext = DSLContextFactory.create(dataSource, SQLDialect.POSTGRES);
-    flyway = FlywayFactory.create(dataSource, DatabaseConfigPersistenceTest.class.getName(), ConfigsDatabaseMigrator.DB_IDENTIFIER,
-        ConfigsDatabaseMigrator.MIGRATION_FILE_LOCATION);
-    database = new ConfigsDatabaseTestProvider(dslContext, flyway).create(false);
     configPersistence = spy(new DatabaseConfigPersistence(database));
-    final ConfigsDatabaseMigrator configsDatabaseMigrator =
-        new ConfigsDatabaseMigrator(database, flyway);
-    final DevDatabaseMigrator devDatabaseMigrator = new DevDatabaseMigrator(configsDatabaseMigrator);
-    MigrationDevHelper.runLastMigration(devDatabaseMigrator);
     truncateAllTables();
-  }
-
-  @AfterEach
-  void tearDown() throws Exception {
-    dslContext.close();
-    DataSourceFactory.close(dataSource);
   }
 
   @Test
@@ -151,7 +202,7 @@ class DatabaseConfigPersistenceTest extends BaseDatabaseConfigPersistenceTest {
   void testGetConnectorRepositoryToInfoMap() throws Exception {
     final String connectorRepository = "airbyte/duplicated-connector";
     final String oldVersion = "0.1.10";
-    final String newVersion = "0.2.0";
+    final String newVersion = DEFAULT_PROTOCOL_VERSION;
     final StandardSourceDefinition source1 = new StandardSourceDefinition()
         .withSourceDefinitionId(UUID.randomUUID())
         .withName("source-1")
@@ -207,13 +258,13 @@ class DatabaseConfigPersistenceTest extends BaseDatabaseConfigPersistenceTest {
 
   @Test
   void testHasNewVersion() {
-    assertTrue(DatabaseConfigPersistence.hasNewVersion("0.1.99", "0.2.0"));
+    assertTrue(DatabaseConfigPersistence.hasNewVersion("0.1.99", DEFAULT_PROTOCOL_VERSION));
     assertFalse(DatabaseConfigPersistence.hasNewVersion("invalid_version", "0.1.2"));
   }
 
   @Test
   void testHasNewPatchVersion() {
-    assertFalse(DatabaseConfigPersistence.hasNewPatchVersion("0.1.99", "0.2.0"));
+    assertFalse(DatabaseConfigPersistence.hasNewPatchVersion("0.1.99", DEFAULT_PROTOCOL_VERSION));
     assertFalse(DatabaseConfigPersistence.hasNewPatchVersion("invalid_version", "0.3.1"));
     assertTrue(DatabaseConfigPersistence.hasNewPatchVersion("0.1.0", "0.1.3"));
   }
@@ -279,6 +330,99 @@ class DatabaseConfigPersistenceTest extends BaseDatabaseConfigPersistenceTest {
         STANDARD_DESTINATION_DEFINITION);
 
     assertThat(filteredSourceMap).containsOnlyKeys(nonCustomKey);
+  }
+
+  void writeSource(final ConfigPersistence configPersistence, final StandardSourceDefinition source) throws Exception {
+    configPersistence.writeConfig(ConfigSchema.STANDARD_SOURCE_DEFINITION, source.getSourceDefinitionId().toString(), source);
+  }
+
+  void writeSourceWithSourceConnection(final ConfigPersistence configPersistence, final StandardSourceDefinition source)
+      throws Exception {
+    configPersistence.writeConfig(ConfigSchema.STANDARD_SOURCE_DEFINITION, source.getSourceDefinitionId().toString(), source);
+    final UUID connectionId = UUID.randomUUID();
+    final UUID workspaceId = UUID.randomUUID();
+    final StandardWorkspace workspace = new StandardWorkspace()
+        .withWorkspaceId(workspaceId)
+        .withName(CANNOT_BE_NULL)
+        .withSlug(CANNOT_BE_NULL)
+        .withInitialSetupComplete(true)
+        .withDefaultGeography(Geography.AUTO);
+    configPersistence.writeConfig(ConfigSchema.STANDARD_WORKSPACE, workspaceId.toString(), workspace);
+
+    final SourceConnection sourceConnection = new SourceConnection()
+        .withSourceId(connectionId)
+        .withWorkspaceId(workspaceId)
+        .withName(CANNOT_BE_NULL)
+        .withSourceDefinitionId(source.getSourceDefinitionId());
+    configPersistence.writeConfig(ConfigSchema.SOURCE_CONNECTION, connectionId.toString(), sourceConnection);
+  }
+
+  void writeDestination(final ConfigPersistence configPersistence, final StandardDestinationDefinition destination)
+      throws Exception {
+    configPersistence.writeConfig(ConfigSchema.STANDARD_DESTINATION_DEFINITION, destination.getDestinationDefinitionId().toString(), destination);
+  }
+
+  void writeDestinationWithDestinationConnection(final ConfigPersistence configPersistence,
+                                                 final StandardDestinationDefinition destination)
+      throws Exception {
+    configPersistence.writeConfig(ConfigSchema.STANDARD_DESTINATION_DEFINITION, destination.getDestinationDefinitionId().toString(), destination);
+    final UUID connectionId = UUID.randomUUID();
+    final UUID workspaceId = UUID.randomUUID();
+    final StandardWorkspace workspace = new StandardWorkspace()
+        .withWorkspaceId(workspaceId)
+        .withName(CANNOT_BE_NULL)
+        .withSlug(CANNOT_BE_NULL)
+        .withInitialSetupComplete(true)
+        .withDefaultGeography(Geography.AUTO);
+    configPersistence.writeConfig(ConfigSchema.STANDARD_WORKSPACE, workspaceId.toString(), workspace);
+
+    final DestinationConnection destinationConnection = new DestinationConnection()
+        .withDestinationId(connectionId)
+        .withWorkspaceId(workspaceId)
+        .withName(CANNOT_BE_NULL)
+        .withDestinationDefinitionId(destination.getDestinationDefinitionId());
+    configPersistence.writeConfig(ConfigSchema.DESTINATION_CONNECTION, connectionId.toString(), destinationConnection);
+  }
+
+  void writeDestinations(final ConfigPersistence configPersistence, final List<StandardDestinationDefinition> destinations)
+      throws Exception {
+    final Map<String, StandardDestinationDefinition> destinationsByID = destinations.stream()
+        .collect(Collectors.toMap(destinationDefinition -> destinationDefinition.getDestinationDefinitionId().toString(), Function.identity()));
+    configPersistence.writeConfigs(ConfigSchema.STANDARD_DESTINATION_DEFINITION, destinationsByID);
+  }
+
+  void deleteDestination(final ConfigPersistence configPersistence, final StandardDestinationDefinition destination)
+      throws Exception {
+    configPersistence.deleteConfig(ConfigSchema.STANDARD_DESTINATION_DEFINITION, destination.getDestinationDefinitionId().toString());
+  }
+
+  protected Map<String, Set<JsonNode>> getMapWithSet(final Map<String, Stream<JsonNode>> input) {
+    return input.entrySet().stream().collect(Collectors.toMap(
+        Entry::getKey,
+        e -> e.getValue().collect(Collectors.toSet())));
+  }
+
+  // assertEquals cannot correctly check the equality of two maps with stream values,
+  // so streams are converted to sets before being compared.
+  protected void assertSameConfigDump(final Map<String, Stream<JsonNode>> expected, final Map<String, Stream<JsonNode>> actual) {
+    assertEquals(getMapWithSet(expected), getMapWithSet(actual));
+  }
+
+  protected void assertRecordCount(final int expectedCount, final Table table) throws Exception {
+    final Result<Record1<Integer>> recordCount = database.query(ctx -> ctx.select(count(asterisk())).from(table).fetch());
+    assertEquals(expectedCount, recordCount.get(0).value1());
+  }
+
+  protected void assertHasSource(final StandardSourceDefinition source) throws Exception {
+    assertEquals(source, configPersistence
+        .getConfig(ConfigSchema.STANDARD_SOURCE_DEFINITION, source.getSourceDefinitionId().toString(),
+            StandardSourceDefinition.class));
+  }
+
+  protected void assertHasDestination(final StandardDestinationDefinition destination) throws Exception {
+    assertEquals(destination, configPersistence
+        .getConfig(ConfigSchema.STANDARD_DESTINATION_DEFINITION, destination.getDestinationDefinitionId().toString(),
+            StandardDestinationDefinition.class));
   }
 
 }
