+++ airbyte-config/config-persistence/src/test/java/io/airbyte/config/persistence/DatabaseConfigPersistenceUpdateConnectorDefinitionsTest.java
@@ -5,32 +5,29 @@
 package io.airbyte.config.persistence;
 
 import static io.airbyte.db.instance.configs.jooq.generated.Tables.ACTOR_DEFINITION;
+import static org.jooq.impl.DSL.asterisk;
+import static org.jooq.impl.DSL.count;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.config.ConfigSchema;
 import io.airbyte.config.StandardSourceDefinition;
+import io.airbyte.config.StandardSourceDefinition.SourceType;
 import io.airbyte.config.persistence.DatabaseConfigPersistence.ConnectorInfo;
-import io.airbyte.db.factory.DSLContextFactory;
-import io.airbyte.db.factory.DataSourceFactory;
-import io.airbyte.db.factory.FlywayFactory;
-import io.airbyte.db.instance.configs.ConfigsDatabaseMigrator;
-import io.airbyte.db.instance.configs.ConfigsDatabaseTestProvider;
-import io.airbyte.db.instance.development.DevDatabaseMigrator;
-import io.airbyte.db.instance.development.MigrationDevHelper;
 import io.airbyte.protocol.models.ConnectorSpecification;
-import io.airbyte.test.utils.DatabaseConnectionHelper;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.UUID;
 import java.util.stream.Collectors;
-import org.jooq.SQLDialect;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
+import org.jooq.Record1;
+import org.jooq.Result;
+import org.jooq.Table;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -39,37 +36,31 @@ import org.junit.jupiter.api.Test;
  * Unit test for the {@link DatabaseConfigPersistence#updateConnectorDefinitions} method.
  */
 @SuppressWarnings("PMD.SignatureDeclareThrowsException")
-class DatabaseConfigPersistenceUpdateConnectorDefinitionsTest extends BaseDatabaseConfigPersistenceTest {
+class DatabaseConfigPersistenceUpdateConnectorDefinitionsTest extends BaseConfigDatabaseTest {
+
+  protected static final StandardSourceDefinition SOURCE_GITHUB = new StandardSourceDefinition()
+      .withName("GitHub")
+      .withSourceDefinitionId(UUID.fromString("ef69ef6e-aa7f-4af1-a01d-ef775033524e"))
+      .withDockerRepository("airbyte/source-github")
+      .withDockerImageTag("0.2.3")
+      .withDocumentationUrl("https://docs.airbyte.io/integrations/sources/github")
+      .withIcon("github.svg")
+      .withSourceType(SourceType.API)
+      .withProtocolVersion("0.2.0")
+      .withTombstone(false);
 
   private static final JsonNode SOURCE_GITHUB_JSON = Jsons.jsonNode(SOURCE_GITHUB);
   private static final String DOCKER_IMAGE_TAG = "0.0.0";
   private static final String DOCKER_IMAGE_TAG_2 = "0.1000.0";
   private static final String DEFAULT_PROTOCOL_VERSION = "0.2.0";
 
-  @BeforeAll
-  public static void setup() throws Exception {
-    dataSource = DatabaseConnectionHelper.createDataSource(container);
-    dslContext = DSLContextFactory.create(dataSource, SQLDialect.POSTGRES);
-    flyway = FlywayFactory.create(dataSource, DatabaseConfigPersistenceUpdateConnectorDefinitionsTest.class.getName(),
-        ConfigsDatabaseMigrator.DB_IDENTIFIER,
-        ConfigsDatabaseMigrator.MIGRATION_FILE_LOCATION);
-    database = new ConfigsDatabaseTestProvider(dslContext, flyway).create(false);
-    configPersistence = new DatabaseConfigPersistence(database);
-    final ConfigsDatabaseMigrator configsDatabaseMigrator =
-        new ConfigsDatabaseMigrator(database, flyway);
-    final DevDatabaseMigrator devDatabaseMigrator = new DevDatabaseMigrator(configsDatabaseMigrator);
-    MigrationDevHelper.runLastMigration(devDatabaseMigrator);
-  }
-
-  @AfterAll
-  public static void tearDown() throws Exception {
-    dslContext.close();
-    DataSourceFactory.close(dataSource);
-  }
+  private DatabaseConfigPersistence configPersistence;
 
   @BeforeEach
   public void resetDatabase() throws SQLException {
     truncateAllTables();
+
+    configPersistence = new DatabaseConfigPersistence(database);
   }
 
   @Test
@@ -225,6 +216,21 @@ class DatabaseConfigPersistenceUpdateConnectorDefinitionsTest extends BaseDataba
     for (final StandardSourceDefinition source : expectedUpdatedSources) {
       assertHasSource(source);
     }
+  }
+
+  void writeSource(final ConfigPersistence configPersistence, final StandardSourceDefinition source) throws Exception {
+    configPersistence.writeConfig(ConfigSchema.STANDARD_SOURCE_DEFINITION, source.getSourceDefinitionId().toString(), source);
+  }
+
+  protected void assertRecordCount(final int expectedCount, final Table table) throws Exception {
+    final Result<Record1<Integer>> recordCount = database.query(ctx -> ctx.select(count(asterisk())).from(table).fetch());
+    assertEquals(expectedCount, recordCount.get(0).value1());
+  }
+
+  protected void assertHasSource(final StandardSourceDefinition source) throws Exception {
+    assertEquals(source, configPersistence
+        .getConfig(ConfigSchema.STANDARD_SOURCE_DEFINITION, source.getSourceDefinitionId().toString(),
+            StandardSourceDefinition.class));
   }
 
 }
